kontrol/{foundry,__main__}: factor out foundry_view command similar to foundry_show
