+++ src/kontrol/__main__.py
@@ -8,7 +8,6 @@ from typing import TYPE_CHECKING
 from pyk.cli.pyk import parse_toml_args
 from pyk.cterm.symbolic import CTermSMTError
 from pyk.proof.reachability import APRFailureInfo, APRProof
-from pyk.proof.tui import APRProofViewer
 
 from . import VERSION
 from .cli import _create_argument_parser, generate_options, get_argument_type_setter, get_option_string_destination
@@ -19,7 +18,6 @@ from .foundry import (
     foundry_list,
     foundry_merge_nodes,
     foundry_minimize_proof,
-    foundry_node_printer,
     foundry_refute_node,
     foundry_remove_node,
     foundry_section_edge,
@@ -29,6 +27,7 @@ from .foundry import (
     foundry_state_load,
     foundry_step_node,
     foundry_unrefute_node,
+    foundry_view,
     init_project,
     read_recorded_state_diff,
     read_recorded_state_dump,
@@ -36,14 +35,12 @@ from .foundry import (
 from .hevm import Hevm
 from .kompile import foundry_kompile
 from .prove import foundry_prove
-from .solc import CompilationUnit
 from .utils import _LOG_FORMAT, _rv_blue, _rv_yellow, check_k_version, config_file_path, console, loglevel
 
 if TYPE_CHECKING:
     from pathlib import Path
     from typing import Final, TypeVar
 
-    from pyk.kcfg.tui import KCFGElem
     from pyk.utils import BugReport
 
     from .options import (
@@ -280,18 +277,7 @@ def exec_view_kcfg(options: ViewKcfgOptions) -> None:
     foundry = _load_foundry(
         options.foundry_root, use_hex_encoding=options.use_hex_encoding, add_enum_constraints=options.enum_constraints
     )
-    test_id = foundry.get_test_id(options.test, options.version)
-    contract_name, _ = test_id.split('.')
-    proof = foundry.get_apr_proof(test_id)
-
-    compilation_unit = CompilationUnit.load_build_info(foundry.build_info)
-
-    def _custom_view(elem: KCFGElem) -> Iterable[str]:
-        return foundry.custom_view(contract_name, elem, compilation_unit)
-
-    node_printer = foundry_node_printer(foundry, contract_name, proof)
-    viewer = APRProofViewer(proof, foundry.kevm, node_printer=node_printer, custom_view=_custom_view)
-    viewer.run()
+    foundry_view(foundry, options)
 
 
 def exec_minimize_proof(options: MinimizeProofOptions) -> None:
